requirements.txt
PyYAML
fastapi
uvicorn[standard]

openxlab
